Simulation with real sensors
In this file we have simulated our system with real sensors that we used in our actual system. First we have added libraries of the sensors and then added those sensors in the Proteus schematic file. Then we have upload .hex file of those sensors for each sensor. After all run the project file and see the result.